Use comma-delimited string for external groups in API endpoints
As an example, rather than having `["prefix-group1", "prefix-group2"]`,
it will now have `"prefix-group1,prefix-group2"` as the value for the
external groups API endpoints (both incoming, such as for updates, and
outgoing, such as in the list endpoint's response).

This makes it easier to sync this data with a spreadsheet (e.g. a Google
Sheet).